Remove unnecessary accounting functions
